refactor: Changed tag related endpoints due to the previous commit

+++ src/backend/app/app/crud/tag.py
@@ -1,68 +1,79 @@
 from sqlalchemy.orm import Session
 from fastapi import HTTPException, status
-from typing import Any
+from typing import List
 
 from app.db.models.user_tag import UserTag
 from app.db.models.assign_user import AssignUser
 from app.models.tag import Tag
+from app.crud.user import search_user_by_id
 
-def create_user_tag(*, session: Session, new_tag: Tag) -> Any:
-    tag = search_tag(session=session, name=new_tag.name)
-    if not tag:
+
+def search_all_user_tags(*, session: Session) -> List[UserTag]:
+    return session.query(UserTag).all()
+
+
+def search_user_tag_by_name(*, session: Session, tag_name: str) -> UserTag | None:
+    return session.query(UserTag).filter(tag_name == UserTag.name).first();
+
+
+def create_user_tag(*, session: Session, new_tag: Tag):
+    tag = search_user_tag_by_name(session=session, tag_name= new_tag.name)
+    if tag:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="There is a user tag with this name alredy",
+        )
+    else:
         tag = UserTag(
-            name = new_tag.name,
+            name=new_tag.name
         )
         session.add(tag)
         session.commit()
         session.refresh(tag)
         return tag
-    else:
+
+
+def search_user_tags(*, session: Session, user_id: int) -> List[UserTag]:
+    if not search_user_by_id(session=session, id=user_id):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="There is alredy a tag with this name",
+            detail=f"There is not a user with id {user_id}"
         )
+    else:
+        return session.query(UserTag.name)\
+            .join(AssignUser, UserTag.id == AssignUser.id_user_tag)\
+            .filter(AssignUser.id_user == user_id)\
+            .all()
     
-def get_all_user_tags(*, session: Session) -> Any:
-    return session.query(UserTag).all();
+
+def user_has_tag_assigned(*, session: Session, user_id: int, tag_name: str) -> AssignUser:
+    return session.query(AssignUser)\
+        .join(UserTag)\
+        .filter(AssignUser.id_user == user_id, UserTag.name == tag_name)\
+        .first()
     
-def search_tag(*, session: Session, name: str) -> Any:
-    return session.query(UserTag).filter(UserTag.name == name).first()
 
-def search_current_user_tags(*, session: Session, current_user_id: int):
-    return session.query(AssignUser.tag_name).filter(AssignUser.id_user == current_user_id).all()
+def assign_user_tag(*, session: Session, user_id: int, tag_name: str) -> UserTag | None:
+    tag = search_user_tag_by_name(session=session, tag_name=tag_name)
+    alredy_assigned = user_has_tag_assigned(session=session, user_id=user_id, tag_name=tag_name)
 
-def user_has_tag(*, session: Session, user_id: int, tag_name: str):
-    return session.query(AssignUser).filter(AssignUser.id_user == user_id, AssignUser.tag_name == tag_name).first()
-
-def assign_tags_to_user(*, session: Session, user_id: int, tag_name: str):
-    tag = search_tag(session=session, name=tag_name)
-    alredy_assigned = user_has_tag(session=session, user_id=user_id, tag_name=tag_name)
-    
     if tag and not alredy_assigned:
         user_tag = AssignUser(
             id_user=user_id,
-            tag_name=tag_name,
+            id_user_tag = tag.id,
         )
         session.add(user_tag)
         session.commit()
         session.refresh(user_tag)
-        return user_tag
+        return tag
     
-def remove_tags_from_user(*, session: Session, user_id: int, tag_name: str):
-    tag = search_tag(session=session, name=tag_name)
-    if tag:
-        assigned = user_has_tag(session=session, user_id=user_id, tag_name=tag_name)
-        if assigned:
-            session.delete(assigned)
-            session.commit()
-            return tag_name
-        else:
-            raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="User has not this tag assigned",
-        )
-    else:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="There is no user tag with this name",
-        )
+
+def remove_user_tag(*, session: Session, user_id, tag_name: str) -> UserTag | None: 
+    tag = search_user_tag_by_name(session=session, tag_name=tag_name)
+    assigned = user_has_tag_assigned(session=session, user_id=user_id, tag_name=tag_name)
+
+    if tag and assigned:
+        session.delete(assigned)
+        session.commit()
+        return tag
+
